Fix queued status remaining on optimized builds
If TeamCity removes a queued build in favor of an already existing
build, the queued status will forever remain published by the commit
status publisher.

This adds in some logic to attempt to determine what the correct build
status should be and publish that instead when the build is optimized
out.